tests: fix crash in stream tests
The StreamChannelHelper retriggers a read as soon as the previous read
is finished, but there's a test that closes the channel and it races
with the read loop from StreamChannelHelper.

I made a BlockingChannel to prevent that second read from occurring.
Also, in the failing tests, we don't bridge an encrypt stream with a
decrypt stream because it's harder to test, since decrypt can't start
before encrypt has encrypted a whole chunk.