Tanker: keep instance alive until all calls finish
We have some flaky tests because Tanker instances are GC-ed in the
middle of a test (cannot_attach_an_already_attached_identity
specifically, but it's not relevant).

This happens because the test looks like

```kotlin
    val tankerAlice = Tanker()
    // ...
    tankerAlice.verifyProvisionalIdentity(EmailVerification(aliceEmail, aliceVerificationCode)).get()
    // and never use tankerAlice ever again
```

R8 will optimize this last statement like:

```smali
    invoke-virtual {v3, v7}, Lio/tanker/api/Tanker;->verifyProvisionalIdentity(Lio/tanker/api/Verification;)Lio/tanker/api/TankerFuture;
    move-result-object v3
```

Or in pseudo code:

```java
    // v3 contains Tanker
    Object v3 = new Tanker();
    // ...
    // and now v3 contains the TankerFuture and there is no reference
    // left to Tanker, which can be GC at any time.
    Object v3 = v3.verifyProvisionalIdentity();
```

So this patch implements the keep alive feature on TankerFuture itself
because we need it everywhere anyway. Note that the TankerFuture is
itself kept alive in its companion object's lifeSupport attribute and is
removed from there when the future callback runs.