fix(test): Closing the write stream doesn't guarantee reads will throw
Like in TCP, if the writer closes their half of the socket,
there is no guarantee that the reader will get an EOF or an exception
when they call read().

There can be buffered data in-flight between the closed write side
and the still reading side.
In fact, with encryption sessions, the header in the first block
happens to be buffered, so closing the input side doesn't make
the output side throw immediately.