fix: Forward system props to instrumented tests, not just env vars
Tanker functional tests need the TANKER_ env vars to work,
or alternatively these can be passed as system properties.

The build.gradle was forwarding the env vars, but not the system props.
Android Studio, being a GUI, doesn't read regular shell env vars and
it can be complicated to get it to start with the right Tanker env.

Now we also forward system props, so you can put those env vars
in a gradle.properties in your GRADLE_USER_HOME (~/.gradle)