maint: Upgrade to Gradle version catalogs (new default practice)
This is the new default way to manage dependencies with Gradle.
There's not really any benefit for a single application project
with a single module, which is probably the majority of cases,
but Google has big multi-module projects so a centralized file
for version numbers is the new hot thing.

As a result, there are now 4 files used for declaring dependencies:
- settings.gradle.kts for the repositories
- build.gradle.kts to pre-declare the plugins
- app/build.gradle.kts to actually import the plugins
- gradle/libs.versions.toml to define the actual plugin ID and version