feat: Convert native.gradle to a Kotlin buildSrc, and support macos
Previously native.gradle was a Groovy DSL script, which is notoriously
painful and unintuitive to work with.

Now it's a modern Kotlin DSL script that contains a Gradle task,
and it's cleanly separated in a buildSrc, as recommended by
modern Gradle practices.

Oh, and thanks to this big refactor/migration, it's now really easy
to support building on macOS. That would not have been fun in Groovy.