maint: Remove unused debug/release option of gradle native tasks
Ever since we moved the Conan install part from Gradle to Python
(which is a... very long time ago!), the Gradle task that sets up
the sdk-native libraries just does a copy of a conan package
that was pre-installed by run-ci.

This means Gradle doesn't chose which sdk-native library
gets installed, Gradle cannot choose between debug/release.
And it didn't. The option was just completely ignored.

Now Android also has a concept of debug/release with build variants,
but that's separate from the sdk-native build type, and in practice
we copy the same sdk-native lib to both Android variants.

So by default you have a release sdk-native lib,
even in a Debug android build variant.
It's possible to get a debug sdk-native,
but you have to do that in run-ci.py by adding the debug
conan profile, there's nothing the Gradle task can do about it.